--- client/src/utils/Constants/MoneyMass.ts
@@ -1,9 +1,0 @@
-export const JEDEN_GR = 1.64;
-export const DWA_GR = 2.13;
-export const PIEC_GR = 2.59;
-export const DZIESIEC_GR = 2.51;
-export const DWADZIESCIA_GR = 3.22;
-export const PIECDZIESIAT_GR = 3.94;
-export const JEDEN_ZL = 5;
-export const DWA_ZL = 5.21;
-export const PIEC_ZL = 6.54;
