+++ client/src/utils/Constants/index.ts
@@ -1,4 +1,4 @@
 export * from './ErrorMessages';
 export * from './formsErrorMessages';
-export * from './MoneyMass';
+export * from './Money';
 export * from './storageKeys';
